services/api/requirements.txt
flask
gunicorn
grpcio
protobuf
